backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
openai-whisper
pyttsx3
gtts
pymongo
